pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = [
    { name = "CPE-SMARTCITY\\10100", email = "wang.he@cpe-smartcity.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "output (>=1.0.1,<2.0.0)",
    "cprint (>=1.2.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "django-scheduler (>=0.10.1,<0.11.0)",
    "schedule (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"